Change the type of strlen in SPVM_UTF8_iterate to size_t. #843

+++ lib/SPVM/Builder/include/spvm_utf8.h
@@ -6,7 +6,7 @@
 
 #include "spvm_typedecl.h"
 
-int32_t SPVM_UTF8_iterate(const uint8_t *str, int32_t strlen, int32_t *dst);
+int32_t SPVM_UTF8_iterate(const uint8_t *str, size_t strlen, int32_t *dst);
 
 int32_t SPVM_UTF8_is_valid_utf8_code_point(int32_t code_point);
